STY moved get_turnover to txn.py. Added parameter to set average vs. total turnover.
