+++ pyfolio/txn.py
@@ -119,7 +119,7 @@ def adjust_returns_for_slippage(returns, turnover, slippage_bps):
     turnover: pd.Series
         Time series of daily total of buys and sells
         divided by portfolio value.
-            - See pos.get_turnover.
+            - See txn.get_turnover.
     slippage_bps: int/float
         Basis points of slippage to apply.
 
@@ -173,3 +173,45 @@ def create_txn_profits(transactions):
     profits_dts = pd.DataFrame(txn_descr)
 
     return profits_dts
+
+
+def get_turnover(transactions, positions, period=None, average=True):
+    """
+    Portfolio Turnover Rate:
+
+    Value of purchases and sales divided
+    by the average portfolio value for the period.
+
+    If no period is provided the period is one time step.
+
+    Parameters
+    ----------
+    transactions_df : pd.DataFrame
+        Contains transactions data.
+        - See full explanation in tears.create_full_tear_sheet
+    positions : pd.DataFrame
+        Contains daily position values including cash
+        - See full explanation in tears.create_full_tear_sheet
+    period : str, optional
+        Takes the same arguments as df.resample.
+    average : bool
+        if True, return the average of purchases and sales divided
+        by portfolio value. If False, return the sum of
+        purchases and sales divided by portfolio value.
+
+    Returns
+    -------
+    turnover_rate : pd.Series
+        timeseries of portfolio turnover rates.
+    """
+
+    traded_value = transactions.txn_volume
+    portfolio_value = positions.sum(axis=1)
+    if period is not None:
+        traded_value = traded_value.resample(period, how='sum')
+        portfolio_value = portfolio_value.resample(period, how='mean')
+    # traded_value contains the summed value from buys and sells;
+    # this is divided by 2.0 to get the average of the two.
+    turnover = traded_value / 2.0 if average else traded_value
+    turnover_rate = turnover / portfolio_value
+    return turnover_rate
